#36 updated documentation and moved exportHandler
